tests/wpt/web-platform-tests/websockets/interfaces/WebSocket/events/013.html

<!doctype html>
<title>WebSockets: setting event handlers to {handleEvent:function(){}}</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../constants.js?pipe=sub></script>
<div id=log></div>
<script>
var events = ['onclose', 'onopen', 'onerror', 'onmessage'];
for (var i = 0; i < events.length; ++i) {
  test(function(t) {
    var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
    ws[events[i]] = {handleEvent:function(){}};
    assert_equals(ws[events[i]], null);
  }, events[i]);
};
</script>
